requirements.txt
# Core numerical and scientific libraries
numpy>=1.21.0
cupy-cuda11x>=11.0.0  # Choose appropriate CUDA version for your system
h5py>=3.7.0
pyyaml>=6.0

# Visualization
matplotlib>=3.5.0
open3d>=0.15.0

# Progress bars and logging
tqdm>=4.64.0

# Optional but recommended development tools
pytest>=7.0.0  # For testing
black>=22.0.0  # For code formatting
mypy>=0.950  # For type checking

# Note: Some system-specific dependencies like CUDA toolkit 
# may need to be installed separately for cupy to work 